added observing_window_loader
