+++ app/__init__.py
@@ -31,6 +31,7 @@ db = SQLAlchemy()
 
 from app.dataloader import (
     ObservationLoader,
+    ObservingWindowLoader,
     ProposalLoader,
     BlockLoader,
     InvestigatorLoader,
@@ -39,6 +40,7 @@ from app.dataloader import (
 loaders = {
     "proposal_loader": ProposalLoader(),
     "observation_loader": ObservationLoader(),
+    "observing_window_loader": ObservingWindowLoader(),
     "block_loader": BlockLoader(),
     "investigator_loader": InvestigatorLoader(),
 }
