+++ app/dataloader/observing_window_loader.py
@@ -1,0 +1,126 @@
+import time
+from collections import namedtuple
+import pandas as pd
+import pytz
+from promise import Promise
+from promise.dataloader import DataLoader
+from graphql import GraphQLError
+from app import db
+from app.util import ObservationStatus
+from datetime import datetime
+import pprint
+
+
+ObservingWindowContent = namedtuple(
+  "ObservingWindowContent", ["night_start", "observing_window", "duration", "window_type"]
+)
+
+BlockObservingWindowContent = namedtuple(
+  "BlockObservingWindowContent", ["past_windows", "todays_windows", "remaining_windows"]
+)
+
+
+class ObservingWindowLoader(DataLoader):
+    """
+    Data loader for observing windows.
+
+    """
+
+    START_OF_DAY_HOURS = 6  # 6:00 UT = 8:00 SAST
+
+    def __init__(self):
+        DataLoader.__init__(self, cache=False)
+
+    def start_of_day(self, timestamp, day_start_hour):
+        seconds_until_start_hour = day_start_hour * 3600
+        seconds_per_day = 24 * 3600
+        seconds_since_midnight = timestamp % seconds_per_day
+        if seconds_since_midnight >= seconds_until_start_hour:
+            # time start offset and midnight
+            return (timestamp - seconds_since_midnight) + seconds_until_start_hour
+        else:
+            # time between midnight and start offset
+            return (timestamp - seconds_since_midnight) - seconds_per_day + seconds_until_start_hour
+
+    def batch_load_fn(self, block_ids):
+        return Promise.resolve(self.get_observing_windows(block_ids))
+
+    def get_observing_windows(self, block_ids):
+        # block observing windows query
+        sql_observing_windows = """
+        SELECT Block_Id, UNIX_TIMESTAMP(VisibilityStart) AS VisibilityStart, 
+        UNIX_TIMESTAMP(VisibilityEnd) AS VisibilityEnd, BlockVisibilityWindowType 
+        FROM BlockVisibilityWindow AS bvw
+        JOIN BlockVisibilityWindowType AS bvwt ON bvw.BlockVisibilityWindowType_Id = bvwt.BlockVisibilityWindowType_Id
+        WHERE Block_Id IN %(block_ids)s
+        ORDER BY VisibilityStart DESC
+        """
+
+        df_block_observing_windows = pd.read_sql(sql_observing_windows, con=db.engine, params=dict(block_ids=block_ids))
+
+        # now timestamp
+        now = int(time.time())
+        # the time today began
+        today = self.start_of_day(now, self.START_OF_DAY_HOURS)
+
+        # collect the values
+        values = dict()
+        for b_id in block_ids:
+            past_windows = set()
+            todays_windows = set()
+            remaining_windows = set()
+            for _, row in df_block_observing_windows.iterrows():
+                if b_id == row["Block_Id"]:
+                    visibility_start = row["VisibilityStart"]
+                    visibility_end = row["VisibilityEnd"]
+                    window_type = row["BlockVisibilityWindowType"]
+                    # the time night started
+                    start_of_night = self.start_of_day(visibility_start, self.START_OF_DAY_HOURS)
+                    # observing window details
+                    observing_window_details = ObservingWindowContent(
+                        night_start=datetime.fromtimestamp(start_of_night).strftime("%d %B %Y"),
+                        observing_window="{} - {}".format(
+                            datetime.fromtimestamp(visibility_start).strftime("%H:%M:%S"),
+                            datetime.fromtimestamp(visibility_end).strftime("%H:%M:%S")
+                        ),
+                        duration=visibility_end - visibility_start,
+                        window_type=window_type
+                    )
+                    if start_of_night < today:
+                        past_windows.add(observing_window_details)
+                    elif start_of_night == today:
+                        todays_windows.add(observing_window_details)
+                    elif start_of_night > today:
+                        remaining_windows.add(observing_window_details)
+
+            # collect details of the block
+            values[b_id] = dict(
+                past_windows=sorted(
+                    past_windows,
+                    key=lambda x: datetime.strptime(x[0], "%d %B %Y")
+                ),
+                todays_windows=sorted(
+                    todays_windows,
+                    key=lambda x: datetime.strptime(x[0], "%d %B %Y")
+                ),
+                remaining_windows=sorted(
+                    remaining_windows,
+                    key=lambda x: datetime.strptime(x[0], "%d %B %Y")
+                ),
+            ),
+
+        def get_observing_window_content(block_id):
+            observing_window = values.get(block_id)[0]
+            if not observing_window:
+                raise GraphQLError(
+                    "There is no observing window for the block with id {block_id}".format(
+                      block_id=block_id
+                    )
+                )
+
+            return BlockObservingWindowContent(**observing_window)
+
+        return [
+            get_observing_window_content(block_id)
+            for block_id in block_ids
+        ]
